[fixed sample] split screens on Google Maps and Mapbox
