+++ buildSrc/build.gradle.kts
@@ -16,7 +16,7 @@ repositories {
 }
 
 dependencies {
-    implementation("dev.icerock:mobile-multiplatform:0.5.2")
+    implementation("dev.icerock:mobile-multiplatform:0.5.3")
     implementation("org.jetbrains.kotlin:kotlin-gradle-plugin:1.3.61")
     implementation("com.android.tools.build:gradle:3.5.2")
 }
