+++ sample/android-app/src/main/AndroidManifest.xml
@@ -19,6 +19,8 @@
             </intent-filter>
         </activity>
 
+        <activity android:name=".GoogleMapsActivity" />
+        <activity android:name=".MapboxActivity" />
 
         <!-- TODO: Replace with your API Key from https://developers.google.com/maps/documentation/ios-sdk/ -->
         <meta-data android:name="com.google.android.geo.API_KEY" android:value="YOUR-API-KEY"/>
