+++ sample/android-app/src/main/java/com/icerockdev/app/MainActivity.kt
@@ -4,88 +4,26 @@
 
 package com.icerockdev.app
 
+import android.content.Intent
 import android.os.Bundle
-import androidx.lifecycle.ViewModelProvider
+import androidx.appcompat.app.AppCompatActivity
+import androidx.databinding.DataBindingUtil
 import com.icerockdev.app.databinding.ActivityMainBinding
-import com.icerockdev.library.MapboxViewModel
-import com.mapbox.mapboxsdk.Mapbox
-import com.mapbox.mapboxsdk.maps.Style
-import dev.icerock.moko.geo.LocationTracker
-import dev.icerock.moko.maps.mapbox.MapboxController
-import dev.icerock.moko.mvvm.MvvmActivity
-import dev.icerock.moko.mvvm.createViewModelFactory
-import dev.icerock.moko.permissions.PermissionsController
 
-
-class MainActivity : MvvmActivity<ActivityMainBinding, MapboxViewModel>() {
-    override val layoutId: Int = R.layout.activity_main
-    override val viewModelVariableId: Int = BR.viewModel
-    override val viewModelClass: Class<MapboxViewModel> = MapboxViewModel::class.java
-
-    override fun viewModelFactory(): ViewModelProvider.Factory {
-        return createViewModelFactory {
-            MapboxViewModel(
-                locationTracker = LocationTracker(
-                    permissionsController = PermissionsController(
-                        applicationContext = applicationContext
-                    )
-                ),
-                mapsController = MapboxController()
-            ).apply { start() }
-        }
-    }
-
+class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
-        Mapbox.getInstance(applicationContext, "YOUR-ACCESS-TOKEN") // or in the application class
         super.onCreate(savedInstanceState)
-        viewModel.locationTracker.bind(lifecycle, this, supportFragmentManager)
+        val binding: ActivityMainBinding = DataBindingUtil.setContentView(
+            this,
+            R.layout.activity_main
+        )
 
-        binding.map.onCreate(savedInstanceState)
-        binding.map.getMapAsync { mapboxMap ->
-            mapboxMap.setStyle(Style.MAPBOX_STREETS) {
-                viewModel.mapsController.bind(
-                    lifecycle = lifecycle,
-                    context = this,
-                    mapboxMap = mapboxMap,
-                    mapView = binding.map,
-                    style = it
-                )
-            }
+        binding.googleMaps.setOnClickListener {
+            startActivity(Intent(this, GoogleMapsActivity::class.java))
         }
-    }
 
-    override fun onStart() {
-        super.onStart()
-        binding.map.onStart()
-    }
-
-    override fun onResume() {
-        super.onResume()
-        binding.map.onResume()
-    }
-
-    override fun onPause() {
-        super.onPause()
-        binding.map.onPause()
-    }
-
-    override fun onStop() {
-        super.onStop()
-        binding.map.onStop()
-    }
-
-    override fun onSaveInstanceState(outState: Bundle) {
-        super.onSaveInstanceState(outState)
-        binding.map.onSaveInstanceState(outState)
-    }
-
-    override fun onLowMemory() {
-        super.onLowMemory()
-        binding.map.onLowMemory()
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        binding.map.onDestroy()
+        binding.mapbox.setOnClickListener {
+            startActivity(Intent(this, MapboxActivity::class.java))
+        }
     }
 }
