+++ sample/mpp-library/src/commonMain/kotlin/com/icerockdev/library/GoogleMapViewModel.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
+ * Copyright 2019 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
  */
 
 package com.icerockdev.library
